build: refactor clear_cache.py

+++ scripts/clear_cache.py
@@ -1,23 +1,25 @@
-"""Script to remove cache files and directories."""
-
-import pathlib
 import shutil
+from pathlib import Path
 
 
 def main():
-    cwd = pathlib.Path(".")
+    cwd = Path().cwd()
     file_extensions = ["*.py[co]", ".coverage", ".coverage.*"]
     directories = ["__pycache__", ".pytest_cache", ".ipynb_checkpoints"]
 
     for file_extension in file_extensions:
         for path in cwd.rglob(file_extension):
+            if "venv" in str(path):
+                continue
             path.unlink()
-            print(f"deleted {path}")
+            print(f"delete - {path}")
 
     for directory in directories:
         for path in cwd.rglob(directory):
+            if "venv" in str(path):
+                continue
             shutil.rmtree(path.absolute(), ignore_errors=False)
-            print(f"deleted {path}")
+            print(f"delete - {path}")
 
 
 if __name__ == "__main__":
